pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "x-search-mcp"
version = "0.1.0"
description = "X (Twitter) 検索 MCP サーバー — Grok x_search API 経由でリアルタイムのX投稿を検索・分析"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
x-search-mcp = "x_search_mcp.server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
